pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-transcription-agent"
version = "1.0.0"
description = "An AI agent for transcribing audio files with history management using Groq and Whisper"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["ai", "transcription", "whisper", "groq", "agent", "langchain", "audio"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]

dependencies = [
    "langchain>=0.1.0",
    "langchain-groq>=0.0.1",
    "langchain-community>=0.0.1",
    "openai-whisper>=20231117",
    "soundfile>=0.12.1",
    "librosa>=0.10.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.0",
]
faster = [
    "faster-whisper>=0.10.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/ai-transcription-agent"
Documentation = "https://github.com/yourusername/ai-transcription-agent#readme"
Repository = "https://github.com/yourusername/ai-transcription-agent"
Issues = "https://github.com/yourusername/ai-transcription-agent/issues"

[project.scripts]
transcription-agent = "src.agent:main"

[tool.setuptools]
packages = ["src", "src.tools"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
